refactor: replace legacy rag package with rag2
- Delete legacy rag implementation files (chunk.ts, embed.ts, ingest.ts, query.ts, types.ts)
- Move all rag2 package contents to rag directory
- Keep the improved rag2 implementation as the main rag package
- Update package.json to reflect the new rag package structure

This completes the migration from the old rag implementation to the new improved version.
