+++ packages/rag/src/errors.test.ts
