+++ packages/rag/src/errors.ts
