+++ packages/rag/tsup.config.ts
@@ -4,5 +4,8 @@ export default defineConfig({
 	entry: ["src/index.ts"],
 	format: ["cjs", "esm"],
 	dts: true,
+	clean: true,
 	sourcemap: true,
+	target: "node18",
+	external: ["pg", "openai"],
 });
