--- packages/rag2/package.json
@@ -1,33 +1,0 @@
-{
-	"name": "@giselle-sdk/rag2",
-	"version": "0.1.0",
-	"description": "Simplified RAG system with type safety",
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
-	"scripts": {
-		"build": "tsup",
-		"check-types": "tsc --noEmit",
-		"dev": "tsup --watch",
-		"test": "vitest",
-		"format": "biome check --write ."
-	},
-	"dependencies": {
-		"@ai-sdk/openai": "catalog:",
-		"ai": "catalog:",
-		"pg": "catalog:",
-		"pgvector": "catalog:",
-		"zod": "catalog:"
-	},
-	"devDependencies": {
-		"@types/node": "catalog:",
-		"@types/pg": "catalog:",
-		"tsup": "catalog:",
-		"typescript": "catalog:",
-		"vitest": "catalog:"
-	},
-	"peerDependencies": {
-		"openai": "catalog:"
-	},
-	"keywords": ["rag", "vector-search", "embeddings", "typescript"],
-	"license": "MIT"
-}
