--- packages/rag2/tsconfig.json
@@ -1,9 +1,0 @@
-{
-	"extends": "../../tools/tsconfig/ts-library.json",
-	"compilerOptions": {
-		"outDir": "./dist",
-		"rootDir": "./src"
-	},
-	"include": ["src/**/*"],
-	"exclude": ["dist", "node_modules", "**/*.test.ts"]
-}
