--- packages/rag2/tsup.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-	entry: ["src/index.ts"],
-	format: ["cjs", "esm"],
-	dts: true,
-	clean: true,
-	sourcemap: true,
-	target: "node18",
-	external: ["pg", "openai"],
-});
