--- packages/rag2/turbo.json
@@ -1,9 +1,0 @@
-{
-	"extends": ["//"],
-	"tasks": {
-		"build": {
-			"outputs": ["**/dist/**"]
-		},
-		"test": {}
-	}
-}
